March 14 tests

+++ UFEC23/Core/Inc/ParticlesManager.h
@@ -21,7 +21,7 @@ typedef struct MeasureParticles
 	uint16_t variance;
 	uint16_t temperature;
 	uint16_t zero;
-	uint16_t normalized_zero;
+	float normalized_zero;
 	int slope;
 	uint16_t LED_current_meas;
 	uint16_t Lux_ON;
@@ -40,7 +40,7 @@ uint16_t Particle_getCH1(void);
 uint16_t Particle_getCH0_OFF(void);
 uint16_t Particle_getCH1_OFF(void);
 uint16_t Particle_getCurrent(void);
-uint16_t Particle_getZeroNorm(void);
+float Particle_getZeroNorm(void);
 uint16_t Particle_getTemperature(void);
 uint16_t Particle_getVariance(void);
 uint16_t Particle_getLuxON(void);
